Add Admin Namespace.
Migrate Nooks#edit out into the new namespace.
