+++ app/assets/stylesheets/admin/nooks.scss
@@ -1,0 +1,8 @@
+// Place all the styles related to the admin/nooks controller here.
+// They will automatically be included in application.css.
+// You can use Sass (SCSS) here: http://sass-lang.com/
+  .edit_nook {
+    .carousel {
+      img { width: 100% };
+    }
+  }
