+++ app/helpers/admin/base_helper.rb
@@ -1,0 +1,2 @@
+module Admin::BaseHelper
+end
