+++ app/helpers/admin/nooks_helper.rb
@@ -1,0 +1,2 @@
+module Admin::NooksHelper
+end
