+++ app/views/admin/nooks/edit.html.erb
