+++ app/views/admin/nooks/index.html.slim
@@ -1,0 +1,1 @@
+= render 'admin_main'
