+++ app/views/nooks/_booking_main.html.slim
@@ -1,12 +1,11 @@
-- #TODO: Replace this placeholder conditional.
-= render current_user.try(:admin?) ? 'admin_filters' : 'booking_filters'
+= render 'booking_filters'
 
 - content_for(:main_menu) do
   div.col-xs-12.col-sm-12.col-md-12.col-lg-12#toolbar
     = render 'shared/nav_links'
 
   div.nooks.col-xs-12.col-sm-12.col-md-12.col-lg-12
-    - [Nook.all].flatten.each.with_index do |nook, i|
+    - Nook.all.each.with_index do |nook, i|
       div.nook-item
         div.nook
           = image_tag nook.photos.first ? nook.photos.first : 'placeholder.jpg'
