+++ app/views/nooks/index.html.slim
@@ -1,1 +1,1 @@
-= render current_user.try(:admin?) ? 'admin_main' : 'booking_main'
+= render 'booking_main'
