+++ app/views/shared/_nook_reservations.html.slim
@@ -1,0 +1,26 @@
+- upcoming = nook.reservations.where('start > ?', DateTime.now).limit(10).order('start asc')
+- if upcoming.empty?
+  div.upcoming-reservation.clearfix
+    = t('reservations.no_reservations')
+- else
+  - upcoming.each_with_index do |reservation, i|
+    div.upcoming-reservation.clearfix
+      - if ((i - 1) < 0 || (upcoming[i - 1].start.strftime('%d') != reservation.start.strftime('%d')))
+        div.upcoming-reservation-day
+          = reservation.start.strftime('%B')
+          = ' '
+          = reservation.start.strftime('%d')
+          - if reservation.start.today?
+            = ' - '
+            = t('reservations.today')
+          - if reservation.start.to_date == Date.tomorrow
+            = ' - '
+            = t('reservations.tomorrow')
+      div.upcoming-reservation-user.col-sm-6
+        = reservation.requester.email
+      div.upcoming-reservation-range.col-sm-6.text-right
+        = reservation.start.strftime('%I:%M%p')
+        = '-'
+        = reservation.end.strftime('%I:%M%p')
+      - if (!upcoming[i + 1].nil? && upcoming[i + 1].start.strftime('%d') != reservation.start.strftime('%d'))
+        div.empty-separator
