+++ spec/controllers/admin/base_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Admin::BaseController, type: :controller do
+
+end
