+++ spec/controllers/admin/nooks_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Admin::NooksController, type: :controller do
+
+end
